pass in base and compare datalayer instead
